The autopatch backend is recycled by (dis)connecting slots to execute and started directly with the GUI itself. Autofocus and pipette detection are working again. Replaced all print statements by logging.info. And modified the GUI with hardware connection buttons.
