+++ PatchClamp/patchclamp_widget.py
@@ -1,147 +1,196 @@
 # -*- coding: utf-8 -*-
 """
-Created on Tue Jun  1 10:54:02 2021
+Created on Mon Apr 12 10:37:51 2021
 
 @author: tvdrb
 """
 
-
 import os
 import sys
+import logging
 
 from PyQt5 import QtWidgets
 from PyQt5.QtWidgets import QWidget, QGridLayout, QPushButton, QGroupBox
+import pyqtgraph.exporters
 import pyqtgraph as pg
 
 # Change path so the Widget can be run either independently or as part of Tupolev.
 if __name__ == "__main__":
     os.chdir(os.getcwd() + '\\..')
-from PatchClamp.patchclamp_backend import CameraThread, AutoPatchThread
+from PatchClamp.patchclamp_backend_v2 import CameraThread, AutoPatchThread
 
 
 class PatchClampUI(QWidget):
-    
-    def __init__(self, camera_handle = None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        
+    def __init__(self):
+        super().__init__()
+
         # =====================================================================
-        # --------------------------- Start of GUI ----------------------------
+        # ---------------------------- Start of GUI ---------------------------
         # =====================================================================
-        # ---------------------- General widget settings ----------------------
+        # ----------------------- General widget settings ---------------------
         self.setWindowTitle("Automatic Patchclamp")
+
+        # ------------------ (Dis)onnect hardware container ------------------
+        hardwareContainer = QGroupBox()
+        hardwareLayout = QGridLayout()
         
-        # -------------------------- Camera section ---------------------------
-        displayContainer = QGroupBox()
-        displayContainer.setMinimumSize(1200, 600)
-        displayLayout = QGridLayout()
+        # Button for connecting camera
+        self.connect_camera_button = QPushButton(text="Camera", clicked=self.toggle_connect_camera, checkable=True)
         
-        # Display to project live camera output
+        # Button for connecting micromanipulator
+        self.connect_micromanipulator_button = QPushButton("Micromanipulator", clicked=self.toggle_connect_micromanipulator, checkable=True)
+        
+        # Button for connecting objective motor
+        self.connect_objective_button = QPushButton("Objective", clicked=self.toggle_connect_objective, checkable=True)
+        
+        hardwareLayout.addWidget(self.connect_camera_button, 0, 0, 1, 1)
+        hardwareLayout.addWidget(self.connect_micromanipulator_button, 1, 0, 1, 1)
+        hardwareLayout.addWidget(self.connect_objective_button, 2, 0, 1, 1)
+        hardwareContainer.setLayout(hardwareLayout)
+        
+        # ------------------------ Snapshot container ------------------------
+        snapshotContainer = QGroupBox()
+        snapshotContainer.setMinimumSize(1200, 600)
+        snapshotLayout = QGridLayout()
+
+        # Display to project live camera view
         self.liveWidget = pg.ImageView()
         self.canvaslive = self.liveWidget.getImageItem()
         self.canvaslive.setAutoDownsample(True)
+
         self.liveWidget.ui.roiBtn.hide()
         self.liveWidget.ui.menuBtn.hide()
         self.liveWidget.ui.histogram.hide()
-        displayLayout.addWidget(self.liveWidget, 0, 0, 1, 2)
-        
+        snapshotLayout.addWidget(self.liveWidget, 0, 0, 1, 2)
+
         # Display to project snapshots
         self.snapshotWidget = pg.ImageView()
         self.canvassnap = self.snapshotWidget.getImageItem()
         self.canvassnap.setAutoDownsample(True)
+
         self.snapshotWidget.ui.roiBtn.hide()
         self.snapshotWidget.ui.menuBtn.hide()
         self.snapshotWidget.ui.histogram.hide()
-        displayLayout.addWidget(self.snapshotWidget, 0, 2, 1, 2)
-        
+        snapshotLayout.addWidget(self.snapshotWidget, 0, 2, 1, 2)
+
         # Button for starting camera acquisition. This button is a property of
         # the container because we need its checkable state in a function.
-        self.toggle_live_button = QPushButton("Pause live")
-        self.toggle_live_button.setCheckable(True)
-        self.toggle_live_button.clicked.connect(self.toggle_live)
-        displayLayout.addWidget(self.toggle_live_button, 1, 0, 1, 1)
-        
+        self.request_pause_button = QPushButton(text="Pause live", clicked=self.toggle_live, checkable=True)
+        snapshotLayout.addWidget(self.request_pause_button, 1, 0, 1, 1)
+
         # Button for making a snapshot
-        request_camera_image_button = QPushButton("Snap image")
-        request_camera_image_button.clicked.connect(self.request_snap)
-        displayLayout.addWidget(request_camera_image_button, 1, 1, 1, 1)
-        
-        # Button for automatic focussing a pipette
-        request_autofocus_button = QPushButton("Autofocus pipette")
-        # request_autofocus_button.clicked.connect(self.request_autofocus)
-        displayLayout.addWidget(request_autofocus_button, 1, 2, 1, 1)
-        
-        # Button for detecting pipette tip
-        request_pipette_coords_button = QPushButton("Detect pipette tip")
-        # request_pipette_coords_button.clicked.connect(self.request_pipette_coords)
-        displayLayout.addWidget(request_pipette_coords_button, 1, 3, 1, 1)
-        
-        displayContainer.setLayout(displayLayout)
-        
+        request_camera_image_button = QPushButton(text="Snap image", clicked=self.request_snap)
+        snapshotLayout.addWidget(request_camera_image_button, 1, 1, 1, 1)
+
+        # Button for autofocus
+        request_autofocus_button = QPushButton(text="Autofocus", clicked=self.request_autofocus)
+        snapshotLayout.addWidget(request_autofocus_button, 1, 2, 1, 1)
+
+        # Button for pipette detection
+        request_detection_button = QPushButton(text="Detect pipette", clicked=self.request_detect)
+        snapshotLayout.addWidget(request_detection_button, 1, 3, 1, 1)
+
+        snapshotContainer.setLayout(snapshotLayout)
+
         # -------------------------- Adding to master -------------------------
         master = QGridLayout()
-        master.addWidget(displayContainer, 0, 0, 1, 1)
+        master.addWidget(hardwareContainer, 0, 0, 1, 1)
+        master.addWidget(snapshotContainer, 0, 1, 1, 1)
 
         self.setLayout(master)
+
+        # =====================================================================
         # ---------------------------- End of GUI -----------------------------
-        
-        # =====================================================================
-        # ------------------------- Activate backend --------------------------
         # =====================================================================
         
-        # Initiate backend threads
-        self.camerathread = CameraThread()
-        self.autopatchthread = AutoPatchThread(self.camerathread)
+        # Fire up backend
+        self.autopatch = AutoPatchThread()
         
-        # Connect our GUI display to the signal slots
-        self.camerathread.snapsignal.connect(self.update_canvassnap)
-        self.camerathread.livesignal.connect(self.update_canvaslive)
-
-        # Start threads
-        self.camerathread.start()
-        
-    def update_canvaslive(self, image):
-        # Update the live canvas with every new camera frame
-        self.canvaslive.setImage(image)
-        
-    def update_canvassnap(self, image):
-        # Update the snap canvas with the newest camera frame
-        self.canvassnap.setImage(image)
-    
-    def toggle_live(self):
-        # Request to pause or continue emitting frames from the camera thread
-        if self.toggle_live_button.isChecked():
-            self.camerathread.stop_canvasupdates()
+    def toggle_connect_camera(self):
+        if self.connect_camera_button.isChecked():
+            # Create instance and pass on to backend
+            self.camerathread = CameraThread()
+            self.autopatch.camera_handle = self.camerathread
+            
+            # Connect canvasses to camera signals
+            self.camerathread.livesignal.connect(self.update_canvaslive)
+            self.camerathread.snapsignal.connect(self.update_canvassnap)
+            
+            # Start camera thread
+            self.camerathread.start()
         else:
-            self.camerathread.start_canvasupdates()
+            # Delete camera thread
+            self.camerathread.__del__()
         
-    def request_snap(self):
-        # Request a snapshot from the camera thread
-        self.camerathread.snap()
-        
-    def request_autofocus(self):
-        print("Autofocus pipette")
-        # Starting the autopatch thread with the pipette autofocus functions
-        self.autopatchthread.started.connect(self.autopatchthread.autofocus_pipette)
-        
-    def request_pipette_coords(self):
-        print("Detect pipette tip")
-        # Starting the autopatch thread with the pipette autofocus functions
-        self.autopatchthread.started.connect(self.autopatchthread.detect_pipette_tip)
+    def toggle_connect_micromanipulator(self):
+        # make this a toggle function, function should initiate and connect or
+        # disconnect and delete.
+        if self.connect_micromanipulator_button.isChecked():
+            pass
+        else:
+            pass
+    
+    def toggle_connect_objective(self):
+        # make this a toggle function, function should initiate and connect or
+        # disconnect and delete.
+        if self.connect_objective_button.isChecked():
+            pass
+        else:
+            pass
+    
+    def toggle_live(self, state):
+        # Request to pause or continue emitting frames from the camera thread
+        if self.request_pause_button.isChecked():
+            self.camerathread.livesignal.disconnect()
+        else:
+            self.camerathread.livesignal.connect(self.update_canvaslive)
 
+    def request_snap(self):
+        self.camerathread.snap()
+
+    def request_autofocus(self):
+        logging.info('Request autofocus button pushed')
+        self.autopatch.request("autofocus")
+
+    def request_detect(self):
+        logging.info('Request detection button pushed')
+        self.autopatch.request("detect")
+
+    def update_canvaslive(self, image):
+        self.canvaslive.setImage(image)
+
+    def update_canvassnap(self, image):
+        self.canvassnap.setImage(image)
+        
     def closeEvent(self, event):
         """ Close event
         This method is called when the GUI is shut down. First we need to stop
         the threads that are steill running, then we accept the close event
-        and quit the application.
+        and quit the widget.
         """
-        self.camerathread.__del__()
-        self.autopatchthread.__del__()
-        event.accept()
+        # Disconnect all signals and slots, here or in class?
+        if hasattr(self, 'camerathread'):
+            self.camerathread.__del__()
+            logging.info('Camera thread stopped successfully')
+        if hasattr(self, 'autopatch'):
+            self.autopatch.__del__()
+            logging.info('Autopatch stopped successfully')
         QtWidgets.QApplication.quit()
-        
-        
-if __name__ == "__main__":
+        event.accept()
+        logging.info('Widget closed successfully')
 
+
+if __name__ == "__main__":
+    def start_logger():
+        logging.basicConfig(
+            level=logging.INFO,
+            format="%(asctime)s [%(levelname)s] %(message)s",
+            handlers=[
+                # logging.FileHandler("autopatch.log"),   # uncomment to write to .log
+                logging.StreamHandler()
+            ]
+        )
+    
     def run_app():
         app = QtWidgets.QApplication(sys.argv)
         pg.setConfigOptions(
@@ -150,5 +199,6 @@ if __name__ == "__main__":
         mainwin = PatchClampUI()
         mainwin.show()
         app.exec_()
-
+        
+    start_logger()
     run_app()
